+++ Fovea.Renderer/Primitives/ConstantMedium.cs
@@ -55,7 +55,7 @@ namespace Fovea.Renderer.Primitives
             return true;
         }
 
-        public BoundingBox GetBoundingBox(double t0, double t1)
+        public BoundingBox GetBoundingBox(float t0, float t1)
         {
             return _boundary.GetBoundingBox(t0, t1);
         }
